Fix folder navigation behavior
