+++ src/webclient/tests/index.js
@@ -4,4 +4,5 @@ import AuthenticationService from './api-service.js';
 import Validator from './validator-test.js';
 import StateManager from './state-manager.js';
 import ChangeTitle from './change-title.js';
+import GetFolderAction from './action/get-folder.js';
